+++ spec/spec_helper.rb
@@ -8,7 +8,7 @@ SimpleCov.start do
   add_filter 'spec/'
   add_group 'Client', client_files
   add_group 'Helpers', helper_path
-  minimum_coverage 87 # TODO: bump up as we increase coverage. Goal: 95%
+  minimum_coverage 89 # TODO: bump up as we increase coverage. Goal: 95%
   minimum_coverage_by_file 10 # TODO: bump up as we increase coverage. Goal: 90%
 end
 
@@ -17,5 +17,7 @@ require_relative 'shared_context'
 require_relative 'support/fake_response'
 
 RSpec.configure do |config|
-  # TODO
+  config.before(:each) do
+    ILO_SDK::ENV_VARS.each { |e| ENV[e] = nil } # Clear environment variables
+  end
 end