+++ spec/unit/rest_spec.rb
@@ -29,7 +29,7 @@ RSpec.describe ILO_SDK::Client do
       expect(Net::HTTP).to receive(:new).with(anything, anything, nil, nil).and_raise('Fake Error')
       expect { @client.rest_api(:get, path) }.to raise_error('Fake Error')
     end
-    
+
     it 'does not disable the proxy unless the disable_proxy client option is set' do
       expect(Net::HTTP).to receive(:new).with(anything, anything).and_raise('Fake Error')
       expect { @client.rest_api(:get, path) }.to raise_error('Fake Error')
@@ -100,6 +100,11 @@ RSpec.describe ILO_SDK::Client do
   describe '#response_handler' do
     it 'returns the JSON-parsed body for 200 status' do
       expect(@client.response_handler(FakeResponse.new(data))).to eq(data)
+    end
+
+    it 'returns the body if it cannot be JSON-parsed for 200 status' do
+      expect_any_instance_of(Logger).to receive(:warn).with(/Failed to parse JSON response/).and_return(true)
+      expect(@client.response_handler(FakeResponse.new('Fake JSON'))).to eq('Fake JSON')
     end
 
     it 'returns the JSON-parsed body for 201 status' do